refactor: コンポーネント群を整理した
